// website/src/notifications_old/types.ts
export interface ToasterStore {
  rendered: Toast[];
  queued: Toast[];
  toasterConfig: Config;
  isWindowBlurred: boolean;
}

export interface Toaster {
  toasterId?: string;
  store: ToasterStore;
  setStore: SetStoreFunction<ToasterStore>;
  counter: number;
}

export type RegisteredToaster = Omit<Toaster, "toasterId"> & {
  toasterId: string;
};

export type ToastType =
  | "default"
  | "success"
  | "error"
  | "loading"
  | "warning"
  | "info";

export interface ToastProviderOptions extends ToasterOptions {
  children: JSX.Element;
}

export type ToasterOptions = Partial<
  Omit<Config, "content" | "contentType" | "id">
>;
// ...
export type ToastContent =
  | string
  | JSX.Element
  | ((toast: Toast) => JSX.Element | string);

export interface ToastPromiseMessages {
  pending: string | JSX.Element;
  success: string | JSX.Element | ((data: any) => string | JSX.Element);
  error: string | JSX.Element | ((error: any) => string | JSX.Element);
}

export interface ToasterContextType {
  toasters: Map<string, Toaster>;
  providerProps: Config;
  registerToaster: (args: Toaster) => RegisteredToaster;
  getToaster: (id?: string) => Toaster;
  unregisterToaster: (id: string) => void;
}
// ...
export interface Config {
  id?: string;
  content?: ToastContent;
  contentType?: "static" | "dynamic";
  theme: "light" | "dark" | string | undefined | null;
  toasterId?: string;
  type: ToastType;
  duration: number | false;
  onEnter: string | undefined;
  enterDuration: number;
  onExit: string | undefined;
  exitDuration: number;
  onIdle: string | undefined;
  positionY: "top" | "bottom";
  positionX: "left" | "right" | "center";
  offsetX: number;
  offsetY: number;
  gutter: number;
  limit: number | false;
  reverseToastOrder: boolean;
  pauseOnHover: boolean;
  pauseOnWindowInactive: boolean;
  renderOnWindowInactive: boolean;
  class: string | ((type: ToastType) => string);
  style:
    | JSX.CSSProperties
    | ((type: ToastType) => JSX.CSSProperties)
    | undefined;
  wrapperClass: string | ((type: ToastType) => string);
  wrapperStyle:
    | JSX.CSSProperties
    | ((type: ToastType) => JSX.CSSProperties)
    | undefined;
  enterCallback: (() => void) | null;
  updateCallback: (() => void) | null;
  exitCallback: ((reason?: boolean | string) => void) | null;
  showDismissButton: boolean;
  dismissButtonClass: string | ((type: ToastType) => string);
  dismissButtonStyle:
    | JSX.CSSProperties
    | ((type: ToastType) => JSX.CSSProperties)
    | undefined;
  dismissOnClick: boolean;
  toasterStyle: Omit<
    JSX.CSSProperties,
    | "top"
    | "bottom"
    | "left"
    | "right"
    | "justify-content"
    | "align-items"
    | "pointer-events"
  > | null;
  showProgressBar: boolean;
  progressBarClass: string | ((type: ToastType) => string);
  progressBarStyle:
    | JSX.CSSProperties
    | ((type: ToastType) => JSX.CSSProperties)
    | undefined;
  showIcon: boolean;
  icon: ((type: ToastType) => JSX.Element) | JSX.Element | null;
  dragToDismiss: boolean;
  dragTreshold: number;
  ariaLive: "off" | "assertive" | "polite";
  role: "status" | "alert";
}